# module_2/openapi.yaml
openapi: 3.1.0
info:
  title: Mini Kanban API
  version: 0.1.0
  description: |
    HTTP contract expected by the frontend `KanbanService` boundary
    (`frontend/src/services/kanban-service.ts` + `types.ts`), plus auth.

    Authentication:
      - `POST /api/auth/login` is public.
      - All board / project / card / settings endpoints require a Bearer token.

    Domain:
      - Board: aggregate of projects, cards, and week settings (primary read)
      - Project: named container; `colorIndex` drives card accent color
      - Card: title + status; belongs to one project; at most one timetable slot
      - Slot: `day` 0–6 (Mon–Sun), `hour` 6–23 (one-hour cell)
      - Settings: optional Monday start date (display labels only)

servers:
  - url: http://localhost:8000
    description: Local FastAPI

tags:
  - name: auth
  - name: board
  - name: projects
  - name: cards
  - name: settings

security:
  - bearerAuth: []

paths:
  /api/auth/login:
    post:
      tags: [auth]
      summary: Log in and receive a bearer token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/board:
    get:
      tags: [board]
      summary: Get full board
      description: |
        Corresponds to `KanbanService.getBoard()`.
        Returns all projects, all cards (with optional slots), and week settings.
      operationId: getBoard
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/projects:
    post:
      tags: [projects]
      summary: Create project
      description: Corresponds to `KanbanService.createProject({ name })`.
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/projects/{id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    patch:
      tags: [projects]
      summary: Rename project
      description: Corresponds to `KanbanService.renameProject({ id, name })`.
      operationId: renameProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRename"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [projects]
      summary: Delete project
      description: |
        Corresponds to `KanbanService.deleteProject({ id })`.
        Deletes the project and every card that belongs to it.
      operationId: deleteProject
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/cards:
    post:
      tags: [cards]
      summary: Create card
      description: |
        Corresponds to `KanbanService.createCard({ projectId, title })`.
        New cards start with status `todo` and `slot: null`.
      operationId: createCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/cards/{id}:
    parameters:
      - $ref: "#/components/parameters/CardId"
    patch:
      tags: [cards]
      summary: Update card title and/or status
      description: Corresponds to `KanbanService.updateCard({ id, title?, status? })`.
      operationId: updateCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [cards]
      summary: Delete card
      description: Corresponds to `KanbanService.deleteCard({ id })`.
      operationId: deleteCard
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/cards/{id}/slot:
    parameters:
      - $ref: "#/components/parameters/CardId"
    put:
      tags: [cards]
      summary: Map or unmap a card on the timetable
      description: |
        Corresponds to `KanbanService.mapCard({ id, slot })`.

        - `slot` set: place the card on that day/hour cell (move if already mapped).
        - `slot: null`: unmap the card back to the task grid.

        A cell holds at most one card. If the target cell is occupied, the previous
        occupant is unmapped (`slot` becomes null) and the moved card takes the cell.
        A card occupies at most one cell.
      operationId: mapCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MapCardRequest"
      responses:
        "200":
          description: Mapped or unmapped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/settings/week:
    put:
      tags: [settings]
      summary: Set Monday start date
      description: |
        Corresponds to `KanbanService.setMondayDate({ mondayDate })`.
        Optional display-only week labeling; returns the full board.
      operationId: setMondayDate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetMondayDateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "401":
          $ref: "#/components/responses/Unauthorized"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Token returned by `POST /api/auth/login`

  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Project id (string, as used by the frontend)
    CardId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Card id (string, as used by the frontend)

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1

    TokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]

    CardStatus:
      type: string
      enum: [todo, in_progress, completed]

    Slot:
      type: object
      required: [day, hour]
      properties:
        day:
          type: integer
          minimum: 0
          maximum: 6
          description: 0 = Monday … 6 = Sunday
        hour:
          type: integer
          minimum: 6
          maximum: 23
          description: Hour cell in the 06:00–23:00 grid

    Project:
      type: object
      required: [id, name, colorIndex]
      properties:
        id:
          type: string
        name:
          type: string
        colorIndex:
          type: integer
          minimum: 0
          description: Index into the frontend project color palette

    ProjectCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    ProjectRename:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    Card:
      type: object
      required: [id, projectId, title, status, slot]
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        status:
          $ref: "#/components/schemas/CardStatus"
        slot:
          description: null when the card sits in the task grid
          oneOf:
            - $ref: "#/components/schemas/Slot"
            - type: "null"

    CardCreate:
      type: object
      required: [projectId, title]
      properties:
        projectId:
          type: string
        title:
          type: string
          minLength: 1

    CardUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        status:
          $ref: "#/components/schemas/CardStatus"
      description: At least one of title or status should be provided
      additionalProperties: false

    MapCardRequest:
      type: object
      required: [slot]
      properties:
        slot:
          description: Target cell, or null to unmap
          oneOf:
            - $ref: "#/components/schemas/Slot"
            - type: "null"

    BoardSettings:
      type: object
      required: [mondayDate]
      properties:
        mondayDate:
          type:
            - string
            - "null"
          format: date
          description: ISO date (yyyy-mm-dd) labeling the Monday column; display only

    SetMondayDateRequest:
      type: object
      required: [mondayDate]
      properties:
        mondayDate:
          type:
            - string
            - "null"
          format: date
          description: ISO date (yyyy-mm-dd), or null to clear

    Board:
      type: object
      required: [projects, cards, settings]
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        settings:
          $ref: "#/components/schemas/BoardSettings"